离线消息处理ing

+++ NowChat-Rest/src/main/java/com/ghj/rest/dao/GroupMessageMapper.java
@@ -5,5 +5,5 @@ import com.ghj.rest.model.GroupMessage;
 /**
  * GoupMessageMapper继承基类
  */
-public interface GroupMessageMapper extends MyBatisBaseDao<GroupMessage, Integer> {
+public interface GroupMessageMapper extends MyBatisBaseDao<GroupMessage, Long> {
 }
